Game Update 1.3.1
https://ronmori.itch.io/elementalist/devlog/42966/game-update-elementalist-131

+++ Assets/Scripts/Fireball.cs
@@ -4,10 +4,10 @@ using UnityEngine;
 
 public class Fireball : MonoBehaviour {
 
-    private int _Damage;
+    private float _Damage;
     private float _Duration;
 
-    public void Setup(int damage, float duration)
+    public void Setup(float damage, float duration)
     {
         _Damage = damage;
         _Duration = duration + Time.time;
